requirements.txt
# Core
torch>=2.0
torchvision>=0.15
numpy>=1.21
opencv-python>=4.5.5
matplotlib>=3.5
seaborn>=0.11

# Notebook & utilities
jupyterlab>=3.0
tqdm>=4.64
scikit-learn>=1.0

# Optional: for file I/O and cleaner path handling
pandas>=1.3
imageio>=2.9
pillow~=11.0.0